requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
distlib==0.3.6
Django==4.2.2
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
dulwich==0.21.5
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.10.1
filelock==3.12.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.0.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.3.2
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
webencodings==0.5.1
xattr==0.10.1
zipp==3.15.0
